[refac] Refactor token handling by renaming and reorganizing Token and TokenType classes
