+++ edukit-core/src/main/java/com/edukit/core/auth/facade/AuthFacade.java
@@ -3,9 +3,9 @@ package com.edukit.core.auth.facade;
 import com.edukit.core.auth.enums.AuthCodeType;
 import com.edukit.core.auth.event.MemberSignedUpEvent;
 import com.edukit.core.auth.facade.response.MemberSignUpResponse;
+import com.edukit.core.auth.jwt.dto.Token;
 import com.edukit.core.auth.service.AuthCodeService;
 import com.edukit.core.auth.service.AuthService;
-import com.edukit.core.auth.service.Token;
 import com.edukit.core.auth.service.TokenService;
 import com.edukit.core.member.entity.Member;
 import com.edukit.core.member.service.MemberService;
