+++ edukit-core/src/main/java/com/edukit/core/auth/jwt/JwtValidator.java
@@ -2,6 +2,7 @@ package com.edukit.core.auth.jwt;
 
 import com.edukit.core.auth.exception.AuthErrorCode;
 import com.edukit.core.auth.exception.AuthException;
+import com.edukit.core.auth.jwt.type.TokenType;
 import io.jsonwebtoken.Claims;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
